Moved shadcn UI components inside a ui folder.
