+++ src/components/GameKeep/GameDetail/GameDetails.tsx
@@ -1,7 +1,7 @@
 import React, { useMemo, useState } from "react"
 
-import Button from "components/Button"
-import { DrawerContent, DrawerDescription, DrawerFooter, DrawerHeader, DrawerTitle } from "components/Drawer"
+import Button from "components/ui/Button"
+import { DrawerContent, DrawerDescription, DrawerFooter, DrawerHeader, DrawerTitle } from "components/ui/Drawer"
 import useSession from "hooks/Supabase/useSession"
 import { useAddUserGameMutation } from "services/userGamesApi"
 import { Game } from "types"
