+++ src/components/GameKeep/GameDrawer/GameDrawer.tsx
@@ -3,9 +3,9 @@ import { useEffect, useState } from "react"
 import { useMediaQuery } from "@react-hook/media-query"
 import { skipToken } from "@reduxjs/toolkit/query"
 
-import Button from "components/Button"
-import { Dialog, DialogContent, DialogDescription, DialogHeader, DialogTitle } from "components/Dialog"
-import { Drawer, DrawerContent, DrawerDescription, DrawerFooter, DrawerHeader, DrawerTitle } from "components/Drawer"
+import Button from "components/ui/Button"
+import { Dialog, DialogContent, DialogDescription, DialogHeader, DialogTitle } from "components/ui/Dialog"
+import { Drawer, DrawerContent, DrawerDescription, DrawerFooter, DrawerHeader, DrawerTitle } from "components/ui/Drawer"
 import useSession from "hooks/Supabase/useSession"
 import { useGetGamesByBggGameIdQuery } from "services/gamesApi"
 import { useAddUserGameMutation } from "services/userGamesApi"
