+++ src/components/ui/Command/Command.tsx
@@ -4,7 +4,7 @@ import { type DialogProps } from "@radix-ui/react-dialog"
 import { Command as CommandPrimitive } from "cmdk"
 import { Search } from "lucide-react"
 
-import { Dialog, DialogContent } from "components/Dialog"
+import { Dialog, DialogContent } from "components/ui/Dialog"
 import { cn } from "lib/utils"
 
 //https://github.com/shadcn-ui/ui/issues/120#issuecomment-1571676598
